1. переделка работы функций в макроязыке. теперь все параметры всех функций опциональны и параметров может быть сколько угодно.    лишние параметры вычисляются, но не используются.
2. в callplugin теперь можно передавать сколько угодно параметров. OpenMacroInfo изменена. Count - количество переданных параметров.
   Values - указатель на массив значений параметров.
   плагины, использующие OPEN_FROMMACRO будут работать некорректно.

+++ far/plugins.cpp
@@ -1084,7 +1084,7 @@ int PluginManager::ProcessConsoleInput(ProcessConsoleInputInfo *Info)
 {
 	int nResult = 0;
 
-	for (int i=0; i<PluginsCount; i++)
+	for (size_t i=0; i<PluginsCount; i++)
 	{
 		Plugin *pPlugin = PluginsData[i];
